fix: add antivirus build checks
