+++ services/analysis-engine/tests/conftest.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from importlib.util import module_from_spec, spec_from_file_location
+from pathlib import Path
+from types import ModuleType
+
+
+def load_module(relative_path: str, module_name: str) -> ModuleType:
+    repo_root = Path(__file__).resolve().parents[3]
+    module_path = repo_root / relative_path
+    spec = spec_from_file_location(module_name, module_path)
+    assert spec is not None
+    assert spec.loader is not None
+    module = module_from_spec(spec)
+    spec.loader.exec_module(module)
+    return module
