+++ services/analysis-engine/tests/test_release_packaging.py
@@ -1,19 +1,8 @@
 from __future__ import annotations
 
-from importlib.util import module_from_spec, spec_from_file_location
 from pathlib import Path
-from types import ModuleType
 
-
-def load_module(relative_path: str, module_name: str) -> ModuleType:
-    repo_root = Path(__file__).resolve().parents[3]
-    module_path = repo_root / relative_path
-    spec = spec_from_file_location(module_name, module_path)
-    assert spec is not None
-    assert spec.loader is not None
-    module = module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
+from conftest import load_module
 
 
 def test_release_packaging_includes_architecture_in_artifact_identity(
